googlecloudfunction_source/requirements.txt
functions_framework==3.*
google-cloud-bigquery
google-cloud-storage